Update caselaw.yaml
changed license to "Open Justice Licence".

As per the endpoint "this licence does not permit computational analysis. If you intend to do any programmatic searching in bulk across the Find Case Law records to identify, extract or enrich contents within the records you will need to apply to perform computational analysis. There are no application charges.". This is to protect against concerns within the justice sector of some potential consequences from use of the data.

Strictly speaking the OJL is not open data - as per the Open Definition - as it can't be used "for any purpose" but I'll leave that to open data purists